* 'core' of gitlab.com:TDCRanila/Digital: (27 commits)
  [Project, Core] Minor refractor to 'GameClock' class. # - LARGE_INTERGER is a signed type, so changing the code to use int64 instead. - Member functions that don't need to be a member moved to source file in Detail namespace.
  [Project, CMake] Actually adding Warning=Error and WholeProgramOptimization tags. #
  [Project, Core] Removing old event code, which is no longer used. #
  [Project, Core] Fixed and issue where the empty brackets would cause syntax conflicts/errors with if-else statements in the 'Release' config. #
  [Project, CMake] Made changes to all build configurations and removed the 'Final' configuration. # - Added /MP (multi-processing build) to all configurations. - Removed the 'Final' Configuration as  it did not provide anything unique compared to 'Release'. - 'Debug'is the main development configuration with no optimization and with runtimechecks. - 'PreRelease' is the configuration with speed optimization, but has enhanced debugging information specifically for optimized code. Also retains most debug defines such as logging and assertion. - 'Release' acts as the production configuration with speed optimization, program optimization, Warning=Error, and has most debugging defines disabled.
  [Tools] Adding specific project build script for VS 2022 Toolset. #
  [Project, Git, CMake] Fixing issue where the 'bx' project would not compile in non-debug builds. - Now properly setting the 'BX_CONFIG_DEBUG' define in all (custom) configurations.(PreProcessorDebug, PreRelease, Release) - Updating bgfx.cmake submodule.
  [Project, Core, Utility] Replacing all smart pointer syntax with alliased versions instead for consistency-sake. #
  [Project, Core] Replacing typedefs with 'using' instead. #
  [Project, Core, GameWorld] Adding GameSystem and GameObjects functionality. # - Added 'GameSystem', which handles 'GameObject' spawning, destruction, etc. - Added "GameObject', which inherits from 'Entity'. It has an additional OnCreate and OnDestruct functions for configuration. - 'EntityDataComponent' renamed from 'EntityRegistrationComponent' and now also stores the EntityType of the entity. The EntityType is the type of the templated class when constructing an Entity. - 'ECSUniverse' now also registers entities for quick acess of EntityHandles<->EntityDataComponent, and DUID <->EntityHandles.
  [Project, Utility] Adding simple system to help create (family) type ids for derived types. #
  [Project, Core] Replacing typedefs with 'using' instead. #
  [Project, GameWorld] TransformComponent changes. # - The 'TransformComponent' class is now split up into a simple  Transform' class and and the component part. 'TransformComponent'now simply inherites from Transform to avoid annoying user syntax. - Function to get the transform matrix named more appropiately to 'GetTransformMatrix.'
  [Project, ECS] Default Entity name made a constexpr variable. #
  [Project, CoreSystems, Editor] Editor Related Changes # - 'Stage' OnAttached/OnRemoved no longer are pure virtual. - 'DockerStage'  class split up in 'EditorDocker' and 'EditorSystem'. Where EditorDocker is a stage that owns all editor related things, whereas the EditorSystems owns other elements and has access to ECS related systems. - Moved MainMenuBar code from the 'DockerStage' to its own class.
  [Project, Editor] Renamed 'EditorGameViewport' to 'GameViewport'. #
  [Project, Editor] Moving editor related files to new folders and fixing include paths. # - Moving DockerStage will be 'System' isntead of a 'Stage' class and thus moved to the GameWorld folder with other systems. - General Editor specific code moved to Editor Module folder.
  [Project, Editor] Excisting temporary Editor code changed to be more refined. # - Created 'EditorElementContainer' which is the main way to add and remove different editor elements. 'EditorElement' is an interface class for that. - Refractored docking code in 'DockerStage'. - Moved the game viewport code from 'DockerStage' class to 'EditorGameViewport'.
  [Project, Rendering] For now force the usage of 'nearest-point' texture sampling flag. #
  [Project, Rendering] Making shared BaseRendererSystem class and using RenderTargetDirector. # - Shared code between rendering systems moved to a single class 'BaseRenderSystem'. ECSystems changed to a non-strict Registrar class for the reflection system. - Added option to change the RenderTarget of the 'RenderingSystem' and 'SpriteSystem'.
  ...

+++ Digital/Digital_/Source/Modules/Resource/Image/ImageData.h
@@ -13,9 +13,9 @@ namespace DFW
 
             std::string file_name;
             std::string file_extension;
-            int32 width;
-            int32 height;
-            int32 components_per_pixel;
+            uint16 width;
+            uint16 height;
+            uint8 components_per_pixel;
             size_t data_size;
             uint8* data;
